Make client more flexible for socket selection
+remove redundancy and separate server and client namespaces

+++ src/unix_socket.cpp
@@ -8,11 +8,11 @@
 
 #include "../include/socket.hpp"
 
-namespace echoserver {
+namespace echoserverclient {
 
 int UnixSocket::createAndBind() {
-    listenSocket = socket(AF_UNIX, SOCK_STREAM, 0);
-    if (listenSocket == -1) {
+    socketFd = socket(AF_UNIX, SOCK_STREAM, 0);
+    if (socketFd == INVALID_SOCKET_FD) {
         throw std::runtime_error("Failed to create the Unix domain listening socket");
     }
 
@@ -21,11 +21,11 @@ int UnixSocket::createAndBind() {
     serverAddr.sun_family = AF_UNIX;
     strncpy(serverAddr.sun_path, socketPath.c_str(), sizeof(serverAddr.sun_path));
 
-    if (bind(listenSocket, (struct sockaddr *)&serverAddr, sizeof(serverAddr)) == -1) {
+    if (bind(socketFd, (struct sockaddr *)&serverAddr, sizeof(serverAddr)) == -1) {
         throw std::runtime_error("Failed to bind the Unix domain listening socket");
     }
 
-    return listenSocket;
+    return socketFd;
 }
 
 int UnixSocket::setupNewConnection() {
@@ -34,8 +34,8 @@ int UnixSocket::setupNewConnection() {
 
         socklen_t clientAddrLen = sizeof(clientAddr);
 
-        int clientSocket = accept(listenSocket, (struct sockaddr *)&clientAddr, &clientAddrLen);
-        if (clientSocket == -1) {
+        int clientSocket = accept(socketFd, (struct sockaddr *)&clientAddr, &clientAddrLen);
+        if (clientSocket == INVALID_SOCKET_FD) {
             if (errno == EWOULDBLOCK || errno == EAGAIN) {
                 return clientSocket; // No more incoming connections
             } else {
@@ -57,7 +57,26 @@ void UnixSocket::initOptions(int socket) {
 }
 
 void UnixSocket::destroy() {
-    close(listenSocket);
-    unlink(socketPath.c_str());
+    if (socketFd != INVALID_SOCKET_FD) {
+        close(socketFd);
+        unlink(socketPath.c_str());
+        socketFd = INVALID_SOCKET_FD;
+    }
 }
-} // namespace echoserver
+
+void UnixSocket::connectToServer(const std::string &serverAddress) {
+    socketFd = socket(AF_UNIX, SOCK_STREAM, 0);
+
+    struct sockaddr_un serverSockAddr {};
+
+    serverSockAddr.sun_family = AF_UNIX;
+    strcpy(serverSockAddr.sun_path, serverAddress.c_str());
+    serverSockAddr.sun_path[0] = 0;
+
+    if (connect(socketFd, (struct sockaddr *)&serverSockAddr, sizeof(serverSockAddr)) == -1) {
+        std::cerr << "Failed to connect to the Unix Domain server." << std::endl;
+        destroy();
+        return;
+    }
+}
+} // namespace echoserverclient
